renames utils.js to normalization.js
